+++ apps/console/src/app/api/signin/nonce/route.ts
@@ -1,9 +1,9 @@
 import { NextRequest, NextResponse } from 'next/server';
-import { justaname, Session, tap } from '../../../../../lib';
+import { getJustaname, Session, tap } from '../../../../../lib';
 
 export const GET = async (req: NextRequest): Promise<NextResponse> => {
-  const session = await Session.fromRequest(req)
-  if (!session?.nonce) session.nonce = justaname.signIn.generateNonce()
+  const session = await Session.fromRequest(req);
+  if (!session?.nonce) session.nonce = getJustaname().signIn.generateNonce();
 
-  return tap(new NextResponse(session.nonce), res => session.persist(res))
-}
+  return tap(new NextResponse(session.nonce), (res) => session.persist(res));
+};