+++ packages/@justaname.id/react/src/lib/constants/query/default-query-config.ts
@@ -1,8 +1,0 @@
-import { NetworkMode } from '@tanstack/react-query';
-
-export const defaultQueryConfig = {
-  staleTime: 1000 * 60 * 5,
-  networkMode: 'offlineFirst' as NetworkMode,
-  refetchOnWindowFocus: false,
-  retry: 0,
-};