+++ packages/@justaname.id/react/src/lib/hooks/ens/useAddressEnsNames.ts
@@ -16,6 +16,7 @@ export const buildAddressEnsNames = (
 export interface UseAddressEnsNamesParams {
   address?: Address;
   chainId?: ChainId;
+  enabled?: boolean;
 }
 
 export interface UseAddressEnsNamesResult {
@@ -38,6 +39,7 @@ export const useAddressEnsNames = (
   const queryClient = useQueryClient();
   const { chainId } = useJustaName();
   const _chainId = params?.chainId || chainId;
+  const _enabled = params?.enabled !== undefined ? params.enabled : true;
   const { getRecords } = useRecords();
 
   const getEnsNamesForAddress = async (
@@ -80,7 +82,10 @@ export const useAddressEnsNames = (
         address: params?.address,
       }),
     enabled:
-      Boolean(params?.address) && Boolean(_chainId) && Boolean(ensClient),
+      Boolean(params?.address) &&
+      Boolean(_chainId) &&
+      Boolean(ensClient) &&
+      Boolean(_enabled),
   });
 
   const getEnsNamesForAddressInternal = async (