+++ packages/@justaname.id/react/src/lib/hooks/primaryName/usePrimaryName.ts
@@ -14,6 +14,7 @@ export const buildPrimaryName = (
 export interface UsePrimaryNameParams {
   address?: Address;
   chainId?: ChainId;
+  enabled?: boolean;
 }
 
 export interface UsePrimaryNameResult {
@@ -37,6 +38,7 @@ export const usePrimaryName = (
   params?: UsePrimaryNameParams
 ): UsePrimaryNameResult => {
   const { chainId, justaname } = useJustaName();
+  const _enabled = params?.enabled !== undefined ? params.enabled : true;
   const _chainId = params?.chainId || chainId;
   const { ensClient } = useEnsPublicClient({
     chainId: _chainId,
@@ -84,7 +86,8 @@ export const usePrimaryName = (
       getPrimaryName({
         address: params?.address,
       }),
-    enabled: Boolean(params?.address) && Boolean(ensClient),
+    enabled:
+      Boolean(params?.address) && Boolean(ensClient) && Boolean(_enabled),
   });
 
   const getPrimaryNameInternal = async (