+++ packages/@justaname.id/react/src/lib/hooks/signIn/useEnsAuth.ts
@@ -3,6 +3,7 @@ import { useQuery } from '@tanstack/react-query';
 import { useJustaName } from '../../providers';
 import { ChainId } from '@justaname.id/sdk';
 import { defaultOptions } from '../../query';
+import axios from 'axios';
 
 export const buildEnsAuthKey = (backendUrl: string) => ['ENS_AUTH', backendUrl];
 
@@ -15,6 +16,7 @@ export type EnsAuth<T extends object = {}> = T & {
 export interface UseEnsAuthParams {
   backendUrl?: string;
   currentEnsRoute?: string;
+  enabled?: boolean;
 }
 
 export interface UseEnsAuthReturn<T extends object = {}> {
@@ -30,7 +32,7 @@ export const useEnsAuth = <T extends object = {}>(
   params?: UseEnsAuthParams
 ): UseEnsAuthReturn<T> => {
   const { backendUrl, routes } = useJustaName();
-
+  const _enabled = params?.enabled !== undefined ? params.enabled : true;
   const _backendUrl = useMemo(
     () => params?.backendUrl || backendUrl || '',
     [backendUrl, params?.backendUrl]
@@ -45,22 +47,23 @@ export const useEnsAuth = <T extends object = {}>(
   );
   const query = useQuery({
     ...defaultOptions,
+    retry: 0,
     queryKey: buildEnsAuthKey(_backendUrl),
     queryFn: async () => {
       try {
-        const response = await fetch(currentEnsEndpoint, {
-          method: 'GET',
+        const response = await axios.get(currentEnsEndpoint, {
           headers: {
             'Content-Type': 'application/json',
           },
-          credentials: 'include',
+          withCredentials: true, // Similar to 'credentials: include' in fetch
         });
-        const json = await response.json();
-        return response.status === 200 ? json : null;
+
+        return response.data;
       } catch (e) {
         return null;
       }
     },
+    enabled: Boolean(_enabled),
   });
 
   return {