+++ packages/@justweb3/efp-plugin/src/lib/components/EFPTab/index.tsx
@@ -1,9 +1,8 @@
 import { useFollowing } from '../../hooks/useFollowing';
 import { useFollowers, useStats } from '../../hooks';
-import { FC, Fragment, useRef, useState } from 'react';
+import { FC, useEffect, useMemo, useRef, useState } from 'react';
 import { Flex, LoadingSpinner, P } from '@justweb3/ui';
-import { JustEnsCard } from '@justweb3/widget';
-import { useInfiniteScroll } from '../../hooks/useInfiniteScroll';
+import { JustEnsCard, useInfiniteScroll } from '@justweb3/widget';
 
 export interface EFPTabProps {
   ens: string;
@@ -11,8 +10,8 @@ export interface EFPTabProps {
 
 export const EFPTab: FC<EFPTabProps> = ({ ens }) => {
   const [tab, setTab] = useState<'following' | 'followers'>('following');
-  const followingRef = useRef<HTMLDivElement>(null);
-  const followersRef = useRef<HTMLDivElement>(null);
+  const followingContainerRef = useRef<HTMLDivElement>(null);
+  const followersContainerRef = useRef<HTMLDivElement>(null);
 
   const {
     following,
@@ -24,6 +23,10 @@ export const EFPTab: FC<EFPTabProps> = ({ ens }) => {
     addressOrEns: ens,
   });
 
+  const followingFlat = useMemo(() => {
+    return following?.pages.flatMap((follow) => follow.following);
+  }, [following]);
+
   const {
     followers,
     fetchMoreFollowers,
@@ -34,15 +37,19 @@ export const EFPTab: FC<EFPTabProps> = ({ ens }) => {
     addressOrEns: ens,
   });
 
+  const followersFlat = useMemo(() => {
+    return followers?.pages.flatMap((follow) => follow.followers);
+  }, [followers]);
+
   useInfiniteScroll(
-    followersRef,
+    followersContainerRef,
     fetchMoreFollowers,
     tab === 'followers' && !isFollowersLoading && hasMoreFollowers,
     50
   );
 
   useInfiniteScroll(
-    followingRef,
+    followingContainerRef,
     fetchMoreFollowing,
     tab === 'following' && !isFollowingLoading && hasMoreFollowing,
     50
@@ -51,6 +58,60 @@ export const EFPTab: FC<EFPTabProps> = ({ ens }) => {
   const { stats, isStatsLoading } = useStats({
     addressOrEns: ens,
   });
+
+  const [nbOfCardsPerFollowingRow, setNbOfCardsPerFollowingRow] = useState(0);
+  useEffect(() => {
+    const element = followingContainerRef.current;
+
+    if (!element) return;
+
+    const resizeObserver = new ResizeObserver((entries) => {
+      for (const entry of entries) {
+        if (entry.contentRect) {
+          const width = entry.contentRect.width;
+          let nbOfCardsPerRowTemp = Math.floor(width / 335);
+          if (nbOfCardsPerRowTemp < 1) {
+            nbOfCardsPerRowTemp = 1;
+          }
+          setNbOfCardsPerFollowingRow(nbOfCardsPerRowTemp);
+        }
+      }
+    });
+
+    resizeObserver.observe(element);
+
+    return () => {
+      resizeObserver.unobserve(element); // Cleanup observer on unmount
+    };
+  }, [followingContainerRef, isStatsLoading, tab]);
+
+  const [nbOfCardsPerFollowersRow, setNbOfCardsPerFollowersRow] = useState(0);
+  useEffect(() => {
+    const element = followersContainerRef.current;
+
+    if (!element) return;
+
+    const resizeObserver = new ResizeObserver((entries) => {
+      for (const entry of entries) {
+        if (entry.contentRect) {
+          const width = entry.contentRect.width;
+
+          let nbOfCardsPerRowTemp = Math.floor(width / 335);
+
+          if (nbOfCardsPerRowTemp < 1) {
+            nbOfCardsPerRowTemp = 1;
+          }
+          setNbOfCardsPerFollowersRow(nbOfCardsPerRowTemp);
+        }
+      }
+    });
+
+    resizeObserver.observe(element);
+
+    return () => {
+      resizeObserver.unobserve(element);
+    };
+  }, [followersContainerRef, isStatsLoading, tab]);
 
   if (isStatsLoading) {
     return (
@@ -167,33 +228,62 @@ export const EFPTab: FC<EFPTabProps> = ({ ens }) => {
 
       {tab === 'following' && (
         <Flex
-          direction={'column'}
           gap={'10px'}
           style={{
             overflow: 'auto',
             maxHeight: 'calc(100% - 37px)',
+            overflowX: 'hidden',
           }}
-          ref={followingRef}
+          wrap={'wrap'}
+          ref={followingContainerRef}
         >
-          {following?.pages
-            .flatMap((follow) => follow.following)
-            .map((follow) => (
-              <Fragment key={ens + follow.data}>
-                <JustEnsCard
-                  addressOrEns={follow.data}
-                  chainId={1}
+          {followingFlat?.map((follow) => (
+            <div
+              key={`display-following-members-${ens}-${follow.data}`}
+              style={{
+                minWidth: '335px',
+                flex: 1,
+                width: '100%',
+              }}
+            >
+              <JustEnsCard
+                addressOrEns={follow.data}
+                style={{ width: '100%' }}
+              />
+            </div>
+          ))}
+
+          {followingFlat &&
+            followingFlat?.length > 0 &&
+            Array.from(
+              {
+                length:
+                  nbOfCardsPerFollowingRow -
+                    (followingFlat?.length % nbOfCardsPerFollowingRow) ===
+                  nbOfCardsPerFollowingRow
+                    ? 0
+                    : nbOfCardsPerFollowingRow -
+                      (followingFlat?.length % nbOfCardsPerFollowingRow),
+              },
+              (_, i) => (
+                <div
+                  key={`display-following-members-${ens}-${i}`}
                   style={{
+                    minWidth: '335px',
+                    flex: 1,
                     width: '100%',
                   }}
-                />
-              </Fragment>
-            ))}
-
+                >
+                  <div style={{ width: '100%', height: '100%' }} />
+                </div>
+              )
+            )}
           {isFollowingFetching && hasMoreFollowing && (
             <Flex
               style={{
                 height: '50px',
                 minHeight: '50px',
+                width: '100%',
                 justifyContent: 'center',
                 alignItems: 'center',
                 position: 'relative',
@@ -207,28 +297,56 @@ export const EFPTab: FC<EFPTabProps> = ({ ens }) => {
 
       {tab === 'followers' && (
         <Flex
-          direction={'column'}
           gap={'10px'}
           style={{
             overflow: 'auto',
+            overflowX: 'hidden',
             maxHeight: 'calc(100% - 37px)',
           }}
-          ref={followersRef}
+          wrap={'wrap'}
+          ref={followersContainerRef}
         >
-          {followers?.pages
-            .flatMap((follow) => follow.followers)
-            .map((follow) => (
-              <Fragment key={ens + follow.address}>
-                <JustEnsCard
-                  addressOrEns={follow.address}
-                  chainId={1}
+          {followersFlat?.map((follow) => (
+            <div
+              key={`display-followers-members-${ens}-${follow.address}`}
+              style={{
+                minWidth: '335px',
+                flex: 1,
+                width: '100%',
+              }}
+            >
+              <JustEnsCard
+                addressOrEns={follow.address}
+                style={{ width: '100%' }}
+              />
+            </div>
+          ))}
+
+          {followersFlat &&
+            followersFlat?.length > 0 &&
+            Array.from(
+              {
+                length:
+                  nbOfCardsPerFollowersRow -
+                    (followersFlat?.length % nbOfCardsPerFollowersRow) ===
+                  nbOfCardsPerFollowersRow
+                    ? 0
+                    : nbOfCardsPerFollowersRow -
+                      (followersFlat?.length % nbOfCardsPerFollowersRow),
+              },
+              (_, i) => (
+                <div
+                  key={`display-followers-members-${ens}-${i}`}
                   style={{
+                    minWidth: '335px',
+                    flex: 1,
                     width: '100%',
                   }}
-                />
-              </Fragment>
-            ))}
-
+                >
+                  <div style={{ width: '100%', height: '100%' }} />
+                </div>
+              )
+            )}
           {isFollowersFetching && hasMoreFollowers && (
             <Flex
               style={{
@@ -237,6 +355,7 @@ export const EFPTab: FC<EFPTabProps> = ({ ens }) => {
                 justifyContent: 'center',
                 alignItems: 'center',
                 position: 'relative',
+                width: '100%',
               }}
             >
               <LoadingSpinner color={'var(--justweb3-primary-color)'} />