+++ packages/@justweb3/widget/src/lib/components/Profile/ContentSection/index.tsx
@@ -1,11 +1,16 @@
-import React, { Fragment, useContext, useEffect, useMemo } from 'react';
 import {
   useAccountEnsNames,
   useAccountSubnames,
   useEnsAvatar,
+  useEnsSubnames,
   useMountedAccount,
+  useRecords,
 } from '@justaname.id/react';
-import { SanitizedRecords, SubnameRecordsRoute } from '@justaname.id/sdk';
+import {
+  ChainId,
+  SanitizedRecords,
+  SubnameRecordsRoute,
+} from '@justaname.id/sdk';
 import {
   A,
   Avatar,
@@ -20,15 +25,17 @@ import {
   TabsList,
   TabsTrigger,
 } from '@justweb3/ui';
+import React, { Fragment, useContext, useEffect, useMemo } from 'react';
 import { getChainIcon } from '../../../icons/chain-icons';
 import { getContentHashIcon } from '../../../icons/contentHash-icons';
 import { getTextRecordIcon } from '../../../icons/records-icons';
-import styles from './ContentSection.module.css';
 import { JustaPlugin } from '../../../plugins';
+import { useJustWeb3 } from '../../../providers';
 import { PluginContext } from '../../../providers/PluginProvider';
-import { ProfileSection } from '../ProfileSection';
 import MetadataCard from '../../MetadataCard';
 import MembersSection from '../MembersSection';
+import { ProfileSection } from '../ProfileSection';
+import styles from './ContentSection.module.css';
 
 export interface ContentProps {
   fullSubname?: string;
@@ -53,6 +60,8 @@ const ContentSection: React.FC<ContentProps> = ({
   const { accountSubnames } = useAccountSubnames();
   const { accountEnsNames } = useAccountEnsNames();
   const [tab, setTab] = React.useState('Main');
+  const { openEnsProfile } = useJustWeb3();
+
   const isProfileSelf = useMemo(() => {
     const isEns = accountEnsNames?.map((ens) => ens.ens).includes(fullSubname);
 
@@ -74,9 +83,40 @@ const ContentSection: React.FC<ContentProps> = ({
     chainId,
   ]);
 
+  const { data } = useEnsSubnames({
+    ensDomain: decodeURIComponent(fullSubname),
+    chainId: chainId as ChainId,
+    isClaimed: true,
+    limit: 15,
+    enabled: fullSubname.split('.').length === 2,
+  });
+
   const isProfileCommunity = useMemo(() => {
-    return fullSubname.split('.').length == 2
-  }, [fullSubname])
+    return (
+      data?.pages &&
+      data?.pages
+        .flatMap((subnameData) => subnameData.data)
+        .flatMap((sub) => sub.ens).length > 0
+    );
+  }, [data]);
+
+  const communityName = useMemo(() => {
+    if (fullSubname.split('.').length === 2) return '';
+    return `${fullSubname.split('.')[1]}.${fullSubname.split('.')[2]}`;
+  }, [fullSubname]);
+
+  const { records: communityRecords } = useRecords({
+    ens: communityName,
+    chainId,
+    enabled: !isProfileCommunity,
+  });
+
+  const memberTabName = useMemo(() => {
+    return `Members (${
+      data?.pages?.flatMap((subnameData) => subnameData)[0].pagination
+        .totalCount
+    })`;
+  }, [data]);
 
   const { createPluginApi } = useContext(PluginContext);
 
@@ -87,7 +127,10 @@ const ContentSection: React.FC<ContentProps> = ({
   }, [fullSubname, chainId]);
 
   const hasTabs = useMemo(() => {
-    return plugins.some((plugin) => plugin.components?.ProfileTab) || isProfileCommunity;
+    return (
+      plugins.some((plugin) => plugin.components?.ProfileTab) ||
+      isProfileCommunity
+    );
   }, [plugins, isProfileCommunity]);
 
   const MainTab = (
@@ -265,6 +308,28 @@ const ContentSection: React.FC<ContentProps> = ({
               margin: '0 15px',
             }}
           />
+          {communityName.length > 0 && (
+            <button
+              onClick={() => {
+                openEnsProfile(communityName, chainId);
+              }}
+              className={styles.communityBtn}
+            >
+              <Avatar
+                src={sanitizeEnsImage({
+                  image: communityRecords?.sanitizedRecords.avatar,
+                  name: communityName,
+                  chainId,
+                })}
+                style={{
+                  border: 'none',
+                  padding: 0,
+                }}
+                size={10}
+              />
+              {communityName}
+            </button>
+          )}
           <Flex direction={'row'} justify={'space-between'} align={'center'}>
             <P
               style={{
@@ -383,7 +448,7 @@ const ContentSection: React.FC<ContentProps> = ({
             <TabsList>
               <TabsTrigger value={'Main'}>Main</TabsTrigger>
               {isProfileCommunity && (
-                <TabsTrigger value={'Members'}>Members</TabsTrigger>
+                <TabsTrigger value={'Members'}>{memberTabName}</TabsTrigger>
               )}
               {plugins.map((plugin) => {
                 const component = plugin.components?.ProfileTab;